Update for new position parameter in Spark 3.1.0 RegExpReplace
